Use vthread pool for concurrency
Data-loader was designed for, well, loading data.  However, it is often also
used as a benchmarking tool.  However, it suffers from a problem that limits
its usefulness: core.async/pipeline is designed to preserve order, but it
has a drawback in that it also introduces head-of-line blocking.

Generally, data-loader doesn't care about preserving order.  However, when
used as a benchmark tool, we care about head-of-line blocking.  With this
problem, a single delayed request can hold up the entire line.

Consider a request that ends up blocking with a timeout of 30s.  The queue
will continue to process until the blocked task is the head, and then the queue
will stall.  Items may still buffer behind the stalled request, but only as
deep as the -c concurrency.  Eventually, the stuck request clears, and the
processing continues.  However, from a flow perspective, the client slows
down or even stops completely during this window.

The result is that the load on the system-under-test can ebb and flow
depending on the error rate.  This materializes as a "saw-tooth" graph
for input rate.  However, the real problem is that the tested system has
an effectively reduced input rate, which changes the intent of the
specified concurrency.

This patch fixes this by forgoing order in favor of a pure thread pool
approach.  The client will now keep -c concurrent requests in flight
regardless of whether any number of requests happen to be slow to
respond.

Signed-off-by: Greg Haskins <greg@manetu.com>