Refactor to subcommand-based design and add support for vectors
N.B. This commit introduces a breaking change to the CLI parameters,
necessitating a bump in the version to v3.x.

The main gist is that we now have a subcommand instead of a -m <mode> switch,
and the various subcommands may have subcommand specific options. Some of the
original options are now "global options," while others have moved down to
the subcommand level.

In addition, the input file is now specified with a '-f' global option instead
of a positional argument.

The significant new feature is that we can optionally generate vector embeddings
for each record. This feature is disabled by default, but can be enabled
with the '--vector-provider' option for attribute-writing subcommands such
as 'onboard' and 'load-attributes'.

Two providers are available: 'random' and 'ollama'. The random provider
generates a random vector and is helpful for a low-friction way to create
embedding traffic when the actual vector does not matter. The ollama
provider integrates with the ollama service to generate actual embeddings.

Signed-off-by: Greg Haskins <greg@manetu.com>